+++ example/string_upper.dart
@@ -1,0 +1,9 @@
+import 'package:support/string.dart';
+
+///
+/// Example of using the `upper` extension method on String.
+///
+void main() {
+  print("FOO BAR BAZ".upper()); // Output: "FOO BAR BAZ"
+  print("foO bAr BaZ".upper()); // Output: "FOO BAR BAZ"
+}